requirements.txt
streamlit>=1.32
fastapi>=0.110
uvicorn[standard]>=0.29
pandas>=2.1
numpy>=1.26
pyyaml>=6.0
crewai>=0.28.8
syft>=0.8.0
shap>=0.45.0
lime>=0.2.0
scikit-learn>=1.4
pqcrypto>=0.1.3
kafka-python>=2.0
websockets>=11.0
